feat: rename casparcg-handler-service to amcp-ccg-handler and caspar-cg-handler to osc-ccg-handler
